requirements.txt
openai==1.3.4
python-telegram-bot==20.6
python-dotenv==1.0.0
requests==2.31.0
pinecone-client==2.2.4
pytz==2023.3.post1
dateparser==1.2.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
google-cloud==0.34.0
google-cloud-pubsub==2.18.4
PyYAML==6.0
alpaca-trade-api==3.0.2
googlesearch-python==1.2.3
newspaper3k==0.2.*